removed unused files and renamed others
