+++ client/src/components/CardDialog/DialogItemGroup/DialogItemGroup.tsx
@@ -1,5 +1,5 @@
 import { Grid } from '@material-ui/core';
-import { IDialogItem } from '../../../interface/detail';
+import { IDialogItem } from '../../../interface/DialogItem';
 import DialogItem from '../DialogItem/DialogItem';
 
 type DialogItemGroupProps = {
